size_t 수정, enum 경고 하나 남음

+++ WeekAssignments/String.h
@@ -31,6 +31,6 @@ public:
 	char& operator[](int index);
 	
 	const char* GetString() const;
-	unsigned int GetLength() const;
+	size_t GetLength() const;
 
 };
